move </dom-module> to bottom of file
Per the code samples in the docs, it seems like the style that is being encouraged is to have the dom-module wrap everything, including the <script> tag. If this is good, the my-greeting should also be updated similarly. It's worth noting that before the change to "fix indentation" on this file, all that needed to change was move the dom-module down, as the indentation was already set up for it to be at the bottom.