Remove double page refresh when adding jshint + js tasks
Adding jshint + js tasks triggers a double refresh request, but only the first
one is executed by the browser, and so, it runs most of the time previous
version of the code (transpiled by babel during last refresh). There is a thus
a race between the first browser refresh request and the transpilation of the
code.
This is due to the jshint task asking explicitely for a one time refresh.
Removing it fixes it and there is then only one refresh, once the code is
properly transpiled.