Update dependencies on README.me
Important dependency instructions have been left out from the README.me. The node module `node-gyp` is included in this project; trying to run npm-install in the project will crash some way through unless its dependencies are handled prior-hand. This will result in other modules the project is dependent on not to be installed and users will not be able to use `gulp` in their project, which is an important feature to the intermediate to advanced package.